Delete test directory
This was autogenerated with the project. Chatted with James and he is
okay to move forward with RSpec so we are going to begin writing our
tests in RSpec and remove all these.

P.S. If you are feeling nervous about this commit all of the tests were
just placeholders that didn't place anything under test.